Changed relative file to absolute file
Using relative files causes a failure if you open a file within vim that is not in your current working directory. This can lead to unexpected results if there is an file name identical to the one you are editing in the working directory, and no markdown generated if there is no file with that name in the working directory.